+++ src/mesido/esdl/esdl_mixin.py
@@ -5,7 +5,7 @@ import logging
 import xml.etree.ElementTree as ET  # noqa: N817
 from datetime import timedelta
 from pathlib import Path
-from typing import Any, Dict, Optional
+from typing import Any, Dict, List, Optional
 
 import esdl.esdl_handler
 
@@ -152,6 +152,9 @@ class ESDLMixin(
         self.override_gas_pipe_classes()
 
         self.name_to_esdl_id_map = dict()
+
+        self.__hot_cold_pipe_relations = dict()
+        self.__unrelated_pipes = list()
 
         super().__init__(*args, **kwargs)
 
@@ -499,7 +502,7 @@ class ESDLMixin(
         -------
         Returns true if the pipe is in the supply network thus not ends with "_ret"
         """
-        return True if pipe not in self.cold_pipes else False
+        return pipe in self.hot_to_cold_pipe_map.keys()
 
     def is_cold_pipe(self, pipe: str) -> bool:
         """
@@ -514,7 +517,7 @@ class ESDLMixin(
         -------
         Returns true if the pipe is in the return network thus ends with "_ret"
         """
-        return pipe.endswith("_ret")
+        return pipe in self.hot_to_cold_pipe_map.values()
 
     def hot_to_cold_pipe(self, pipe: str) -> str:
         """
@@ -530,7 +533,7 @@ class ESDLMixin(
         -------
         string with the associated return pipe name.
         """
-        return f"{pipe}_ret"
+        return self.hot_to_cold_pipe_map.get(pipe, None)
 
     def cold_to_hot_pipe(self, pipe: str) -> str:
         """
@@ -546,7 +549,70 @@ class ESDLMixin(
         -------
         string with the associated hot pipe name.
         """
-        return pipe[:-4]
+        return self.cold_to_hot_pipe_map.get(pipe, None)
+
+    def hot_cold_pipe_relations(self):
+        # Backward compatability: ESDL version before v2110 don't have the related attribute
+        esdl_version = self.__energy_system_handler.energy_system.esdlVersion
+        if esdl_version is not None and esdl_version >= "v2110":
+            for asset in self._esdl_assets.values():
+                if asset.asset_type == "Pipe":
+                    related = False
+                    related_asset = asset.attributes.get("related", False)
+                    if related_asset:
+                        assert (
+                            len(related_asset) == 1
+                        ), "Pipes can only have related supply/return pipe"
+                        related = True
+                        if asset.attributes["port"][0].carrier.supplyTemperature:  # hot_pipe
+                            if asset.name not in self.__hot_cold_pipe_relations.keys():
+                                self.__hot_cold_pipe_relations[asset.name] = related_asset[0].name
+                        elif asset.attributes["port"][0].carrier.returnTemperature:  # cold_pipe
+                            if related_asset[0].name not in self.__hot_cold_pipe_relations.keys():
+                                self.__hot_cold_pipe_relations[related_asset[0].name] = asset.name
+                    if not related and asset.name not in self.__unrelated_pipes:
+                        self.__unrelated_pipes.append(asset.name)
+        else:
+            pipes = self.energy_system_components.get("heat_pipe", [])
+            for pipe in pipes:
+                related = False
+                # test if hot_pipe
+                if not pipe.endswith("_ret"):
+                    cold_pipe = f"{pipe}_ret"
+                    if cold_pipe in pipes:
+                        related = True
+                        if pipe not in self.__hot_cold_pipe_relations.keys():
+                            self.__hot_cold_pipe_relations[pipe] = cold_pipe
+                elif pipe.endswith("_ret"):
+                    hot_pipe = pipe[:-4]
+                    if hot_pipe in pipes:
+                        related = True
+                        if hot_pipe not in self.__hot_cold_pipe_relations.keys():
+                            self.__hot_cold_pipe_relations[hot_pipe] = pipe
+                if not related and pipe not in self.__unrelated_pipes:
+                    self.__unrelated_pipes.append(pipe)
+
+    @property
+    def hot_to_cold_pipe_map(self) -> Dict:
+        """
+        This function return a dictionary of hot pipe names mapped to cold pipe names.
+        """
+        return self.__hot_cold_pipe_relations
+
+    @property
+    def cold_to_hot_pipe_map(self) -> Dict:
+        """
+        This function return a dictionary of cold pipe names mapped to hot pipe names.
+        """
+        return dict(
+            zip(self.__hot_cold_pipe_relations.values(), self.__hot_cold_pipe_relations.keys())
+        )
+
+    @property
+    def unrelated_pipes(self) -> List[str]:
+        """This function return a list of pipe names of all the pipes that don't have a related
+        cold/hot pipe."""
+        return self.__unrelated_pipes
 
     def pycml_model(self) -> _ESDLModelBase:
         """
@@ -572,6 +638,7 @@ class ESDLMixin(
         super().read()
         energy_system_components = self.energy_system_components
         esdl_carriers = self.esdl_carriers
+        self.hot_cold_pipe_relations()
         io = self.io
         self.__profile_reader.read_profiles(
             energy_system_components=energy_system_components,