+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ## Changed
 - Minimize TCO objective in the grow_workflow is now only based on capex and opex that can be influenced.
+- Removed the requirement of "_ret" for the return network pipes, for ESDLversion 21.10 and later. The relation between supply and return pipes is now based on the "related" attribute in the esdl.
 
 ## Fixed
 - Bug: Write updated esdl for 2 port heat pump